+++ src/tests/libica_sha256_test.c
@@ -64,7 +64,7 @@ int new_api_sha256_test(void)
 {
 	sha256_context_t sha256_context;
 	int rc = 0;
-	unsigned int i = 0;
+	int i = 0;
 	unsigned char input_data[1000000];
 	unsigned int  output_hash_length = SHA256_HASH_LENGTH;
 	unsigned char output_hash[SHA256_HASH_LENGTH];
@@ -107,7 +107,7 @@ int new_api_sha256_test(void)
 		unsigned int sha_message_part;
 		memset(input_data, 'a', 1024);
 
-		if (i == FIPS_TEST_DATA_SIZE[2])
+		if (i == (int)FIPS_TEST_DATA_SIZE[2])
 			sha_message_part = SHA_MSG_PART_FIRST;
 		else if (i <= 1024)
 			sha_message_part = SHA_MSG_PART_FINAL;
@@ -143,7 +143,7 @@ int new_api_sha256_test(void)
 		unsigned int sha_message_part;
 		memset(input_data, 'a', 64);
 
-		if (i == FIPS_TEST_DATA_SIZE[2])
+		if (i == (int)FIPS_TEST_DATA_SIZE[2])
 			sha_message_part = SHA_MSG_PART_FIRST;
 		else if (i <= 64)
 			sha_message_part = SHA_MSG_PART_FINAL;
